services/verification/verifi/app/requirements.txt
fastapi
uvicorn
google-cloud-storage
pydantic
